requirements.txt
faker==24.0.0
pandas==2.2.0
numpy==1.26.0
google-cloud-bigquery==3.17.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
apache-beam[gcp]==2.54.0
python-dotenv==1.0.0
functions-framework==3.5.0
